NonLogged Component created. Login Forms component created to reuse the UI for all of three screens
